move topKFor into VSM
