+++ src/java/org/apache/cassandra/index/sai/disk/vector/VectorSourceModel.java
@@ -26,6 +26,8 @@ import io.github.jbellis.jvector.vector.VectorSimilarityFunction;
 
 import static io.github.jbellis.jvector.vector.VectorSimilarityFunction.COSINE;
 import static io.github.jbellis.jvector.vector.VectorSimilarityFunction.DOT_PRODUCT;
+import static java.lang.Math.max;
+import static java.lang.Math.pow;
 import static org.apache.cassandra.index.sai.disk.vector.VectorCompression.CompressionType.BINARY_QUANTIZATION;
 import static org.apache.cassandra.index.sai.disk.vector.VectorCompression.CompressionType.PRODUCT_QUANTIZATION;
 
@@ -130,5 +132,43 @@ public enum VectorSourceModel
             return 1.5;
         else
             return 1.0;
+    }
+
+    /**
+     * @param limit the number of results the user asked for
+     * @param cv the compressed vectors being queried.  Null if uncompressed.
+     * @return the topK >= `limit` results to ask the index to search for, forcing
+     * the greedy search deeper into the graph.  This serves two purposes:
+     * 1. Smoothes out the relevance difference between small LIMIT and large
+     * 2. Compensates for using lossily-compressed vectors during the search
+     */
+    public int topKFor(int limit, CompressedVectors cv)
+    {
+        // if the vectors are uncompressed, bump up the limit a bit to start with but decay it rapidly
+        if (cv == null)
+        {
+            var n = max(1.0, 0.979 + 4.021 * pow(limit, -0.761)); // f(1) =  5.0, f(100) = 1.1, f(1000) = 1.0
+            return (int) (n * limit);
+        }
+
+        // Most compressed vectors should be queried at ~2x as much as uncompressed vectors.  (Our compression
+        // is tuned so that this should give us approximately the same recall as using uncompressed.)
+        // Again, we do want this to decay as we go to very large limits.
+        var n = max(1.0, 0.509 + 9.491 * pow(limit, -0.402)); // f(1) = 10.0, f(100) = 2.0, f(1000) = 1.1
+
+        // per-model adjustment on top of the ~2x factor
+        int originalDimension = cv.getOriginalSize() / 8;
+        if (compressionProvider.apply(originalDimension).matches(cv))
+        {
+            n *= overqueryProvider.apply(cv);
+        }
+        else
+        {
+            // we're using an older CV that wasn't created with the currently preferred parameters,
+            // so use the generic defaults instead
+            n *= OTHER.overqueryProvider.apply(cv);
+        }
+
+        return (int) (n * limit);
     }
 }
