+++ test/unit/org/apache/cassandra/index/sai/memory/VectorMemtableIndexTest.java
@@ -51,8 +51,8 @@ import org.apache.cassandra.dht.Range;
 import org.apache.cassandra.index.sai.IndexContext;
 import org.apache.cassandra.index.sai.QueryContext;
 import org.apache.cassandra.index.sai.SAITester;
-import org.apache.cassandra.index.sai.disk.vector.OverqueryUtils;
 import org.apache.cassandra.index.sai.disk.vector.VectorMemtableIndex;
+import org.apache.cassandra.index.sai.disk.vector.VectorSourceModel;
 import org.apache.cassandra.index.sai.plan.Expression;
 import org.apache.cassandra.index.sai.utils.RangeUtil;
 import org.apache.cassandra.index.sai.utils.ScoredPrimaryKey;
@@ -140,7 +140,7 @@ public class VectorMemtableIndexTest extends SAITester
             Set<Integer> foundKeys = new HashSet<>();
             int limit = getRandom().nextIntBetween(1, 100);
 
-            long expectedNumResults = Math.min(OverqueryUtils.topKFor(limit), keysInRange.size());
+            long expectedNumResults = Math.min(VectorSourceModel.OTHER.topKFor(limit, null), keysInRange.size());
 
             try (var iterator = memtableIndex.orderBy(new QueryContext(), expression, keyRange, limit))
             {
